feat: moved to borrowing service & comments
